+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # MacOs DevBox
 http://devbox.ewave.com/
 
+## Release 3.0.2
+- [MacOs] Added installation of PHP which has been removed from default MacOs distribution
+- [Linux/MacOs] Added compatibility with Docker 4.2 and 4.3. As a result log-level option has been removed from 'docker compose' V2 commands due to docker parsing issue.  
+
 ## Release 3.0.1
 - Added fix to work around docker issue with hanging nginx in reserve-proxy container
 - Fixed repeated import of website SSL certificate which resulted in asking for admin privileges by every project starting/stopping with enabled HTTPS